requirements.txt
annotated-types==0.6.0
anyio==3.6.2
async-generator==1.10
attrs==23.2.0
Authlib==1.3.0
bcrypt==4.1.2
bson==0.5.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.0
cycler==0.12.1
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.109.0
fastapi-login==1.9.3
fonttools==4.47.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
kiwisolver==1.4.5
matplotlib==3.8.2
motor==3.3.2
numpy==1.26.3
openai==1.14.3
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.35.1
tqdm==4.66.1
trio==0.21.0
typing_extensions==4.10.0
urllib3==2.1.0
uvicorn==0.29.0
